backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
llama-api-client==0.1.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
asyncpg==0.29.0
websockets==12.0
python-socketio==5.10.0
aiosqlite==0.19.0

# Development
pytest
pytest-asyncio

# OAuth
authlib
httpx 